added repository

+++ TierList/src/main/java/com/example/TierList/model/Utente.java
@@ -17,8 +17,8 @@ import lombok.NoArgsConstructor;
 public class User {
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
-    private int id;
-    
+    private Long id;
+
     private String username;
     private String password;
     private String ruolo;
